0071/feat: change int type by long type for all cost related fields for entities

+++ SpaceBack/src/main/java/com/example/SpaceBack/model/Building.java
@@ -32,13 +32,13 @@ public abstract class Building {
     @Positive
     private int level;
     @PositiveOrZero
-    private int costMetal;
+    private long costMetal;
     @PositiveOrZero
-    private int costCristal;
+    private long costCristal;
     @PositiveOrZero
-    private int costDeuterium;
+    private long costDeuterium;
     @PositiveOrZero
-    private int costEnergy;
+    private long costEnergy;
     @NotBlank
     private String buildingAvatarUrl;
